fix: prevent integer overflow in num_points multiplication
Cast num_points to size_t before multiplying by element size to prevent
unsigned int overflow when num_points * sizeof(element) exceeds UINT_MAX.
This affects RVV and SSE3 implementations that compute byte counts from
num_points.

+++ kernels/volk/volk_32fc_x2_add_32fc.h
@@ -244,7 +244,7 @@ static inline void volk_32fc_x2_add_32fc_rvv(lv_32fc_t* cVector,
     const float* ina = (const float*)aVector;
     const float* inb = (const float*)bVector;
     float* out = (float*)cVector;
-    size_t n = num_points * 2;
+    size_t n = (size_t)num_points * 2;
     for (size_t vl; n > 0; n -= vl, ina += vl, inb += vl, out += vl) {
         vl = __riscv_vsetvl_e32m8(n);
         vfloat32m8_t va = __riscv_vle32_v_f32m8(ina, vl);
